pyproject.toml
[project]
name = "curso-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "langchain>=1.1.0",
    "langchain-classic>=1.0.0",
    "langchain-google-genai>=3.2.0",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
]
